+++ index.android.js
@@ -10,8 +10,6 @@ import {
 } from 'react-native';
 
 import { AppRouter } from './js/router/Router'
-import { eventBus } from './js/util/eventBus'
-import { quitApp } from './js/util/util'
 import { INITIALIZER } from './js/initialize'
 import { colors, screenWidth, screenHeight } from './js/views/styles'
 
@@ -27,25 +25,11 @@ class Root extends Component {
   componentDidMount() {
     // start the BLE things.
     INITIALIZER.init();
-
-    // avoid closing the app by tapping back too often.
-    BackAndroid.addEventListener('hardwareBackPress', () => {
-      Alert.alert("Would you like to close the app?", "The app will still run in the background, this is required for your Crownstones to respond to you. Choose \"Force quit\" to stop running completely.",[
-        {text:'Force quit', onPress: () => { quitApp(); }},
-        {text:'Cancel'},
-        {text:'Ok', onPress: () => { BackAndroid.exitApp(); }}
-      ]);
-
-      // if the user presses back the second time, we close the app
-      return true;
-    });
-
   }
 
   componentWillUnmount() {
     this.unsubscribe.forEach((callback) => {callback()});
     this.unsubscribe = [];
-    this.keyboardDidHideListener.remove();
   }